FP correction & revision for better comprehension
I've fixed some minor errors and revised some texts in order to provide better sense for the text. 
Beside that, I've reduced the text (keeping it's meaning of course) of the buttons close to the bottom of the front page, since on Firefox, Chrome and Safari, the buttons seems to not be adaptive oriented as the main text divs.