Kanban: Move “Add Item” to Header
This patch moves the button to add new items to kanban boards to the
column headers.

The reasoning behind this change is that on larger boards the buttons at
the bottom may be pushed far away from the actual content, causing users
to constantly scroll up and down over the whole page just to add new
items.

This also helps on mobile devices since you may not see all of the
columns there due to the limited screen space. This means that you have
to remember which column you are working on to not accidentally add
items to the wrong column. This patch makes it clear to which  column
users add items since the button now resides next to the column title.