add " (copy)" to title of copied file
This commit basically adds " (copy)" to all copies of files.

The "Make a copy" button sends an event to
`./www/common/sframe-common-outer.js` with a ton of variables. Luckily,
the title of the file is available as originalTitle and all I had to do
was simply add `+ " (copy)"` for it to work.