make buttons visible in mobile version of presentation mode in slides
Slides hides the buttons needed for navigation and exiting in presentation
mode. This makes sense because when doing a presentation, a user would
prefer to not have button icons all over their presentation content. When
a user hovers on the presentation, the needed buttons become visible
again.

This commit makes mobile touch act similarily to hovering over the
presentation.