make mobile version of calendar close date field on 2nd button click
In the Calendar app on mobile, if you click on the calendar date button
it shows a date input field and that input field disappears if you set
a date. If you don't, the input field stays unless you change the date
by clicking on the calendar date button and change the current date.

This fix makes it so that when the calendar date button is clicked and
the input field is visible, it makes the input field invisible.