make contacts page keyboard accessible
The Contacts page is the second previously less accessible component in
issue 1162. This commit hooks all the necessary "keydown" and "keypress"
events and the "tabindex" attribute.

Beyond that, this commit has one change that is important to denote. That is, the sorting mechanism of the contacts page. The sorting mechanism used to be done through the CSS attribute: "order". However, this only changes the visual order not the functional element order. This particularily affects keyboard navigation and makes it completely random.

The new change of this commit makes the sorting method through jQuery's sort function. Other methods simply erased away very important callbacks and other internal DOM properties that erase away much of the functionality. This approach is preferred because it reduces code complexity.