make npm run lint silent + make /teams keyboard accessible
Beside making linter silent and /teams keyboard accessible, there was
a modification to the styling of the messaging elements that allowed
/contacts to be sorted through Javascript.

There was another modification to the styling of the drive to make the
keyboard focus border visible. Lastly, most of the modification done on
./www/teams/inner.js was through this acme command (that uses sed):

Edit , | sed 's|.click(function () {|.on("click keypress", function (e) {\nif(e.type === "keypress" \&\& e.originalEvent.which !== 13) return;|g'

The command above basically replaced all instance of "click(" to an
event handler that subscribes to "keypress" that starts with checking
that the enter key was entered. If not, it just stops the handler.