env.js: fix httpSafePort handling
Env.httpSafePort would only be set if neither httpSafeOrigin nor
httpSafePort are set in config:
 - if neither are set, cryptpad would listen to httpPort+1
 - otherwise there is no way to make it listen to this port

In practice cryptpad does not seem to differentiate what comes on which
port so this looks like purely a dev option (so running from node
directly can bind on two sockets to test different origins), so this
is probably just unused, but the default is to listen to httpPort+1
(3001) and that is what the config describes, so do this.

It would probably make more sense to not listen to this one by default,
as most people would have httpSafeOrigin set, so it is probably not
bound for most people.