Make docker more configurable
At the moment, the docker deployment needs a complete file in a fixed
path (./config/config.js). There was a variable `CPAD_CONF` but it is
not used anymore in the code, I guess it became `CRYPTPAD_CONFIG` but
the docker manifest was never updated.

I think it would be great to get a new interface to configure the
container. For example by passing variables to the container instead of
a big file. I could test and set the environment variable to what looked
the more like what is in config.example.js and made some adjustments
because it is a container.

It also gives a partial fix for #509 .
The remaining volumes mount path are not configurable atm.